% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAxesToRgl.R
\name{addAxesToRgl}
\alias{addAxesToRgl}
\title{Add axis markers at local coordinate space origin}
\usage{
addAxesToRgl(len = 1, sca = NULL, rot = NULL, trn = NULL,
  title = NULL, charexp = 1)
}
\arguments{
\item{len}{The length of each axis}

\item{sca}{scale vector for local coord system (1 coord per axis)}

\item{rot}{rotation matrix  for local coord system}

\item{trn}{translation vector for local coord system (1 coord per axis)}

\item{title}{An optional title}

\item{charexp}{character expansion factor for axis label}
}
\description{
Add axis markers at local coordinate space origin
}
\examples{
addAxesToRgl()
}

