% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw3dlib.R
\docType{package}
\name{mw3dlib}
\alias{mw3dlib}
\alias{mw3dlib-package}
\title{nw3dlib: Mike Wise's package that provides some utility functions for rgl.}
\description{
The mw3dlib package provides utility functions for rgl.
}
\section{mw3dlib functions}{

The mw3dlib functions
    addAxesToRgl() - Adds axes to a plot
    readCompObjCrazyflie() - read in the crazyfile compObj object from csv files
    readCompObj() - read in a compObj object from csv files
    renderCompObj() - render a compObj to rgl
    rgbToStringColor() - convert numeric rgb values to string color specifiers
}

