% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCompObj3d.R
\name{readCompObj3d}
\alias{readCompObj3d}
\title{Read in a compObj object from files}
\usage{
readCompObj3d(objnameroot, objdir = NULL, quiet = T)
}
\arguments{
\item{objnameroot}{the root object name - this forms part of the filenames}

\item{objdir}{the root object directory}

\item{quiet}{A boolean value that supresses reading feedback}
}
\description{
Read in a compObj object from files
}
\examples{
robj <- readCompObj3d("crazyflie")
}

