% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCompObjCrazyflie.R
\name{readCompObjCrazyflie}
\alias{readCompObjCrazyflie}
\title{Read in the Crazyflie files}
\usage{
readCompObjCrazyflie(quiet = T)
}
\arguments{
\item{quiet}{A boolean value that supresses reading feedback}
}
\description{
Read in the Crazyflie files
}
\examples{
robj <- readCompObjCrazyflie()
}

