% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderCompObj3d.R
\name{renderCompObj3d}
\alias{renderCompObj3d}
\title{Render a compObj object in rgl}
\usage{
renderCompObj3d(obj, exclude = NULL, include = NULL, lax = F, quiet = T,
  wireframe = F)
}
\arguments{
\item{obj}{the compobj to render}

\item{exclude}{Layers to exclude}

\item{include}{If layers have been excluded, then these layers will be included anyway}

\item{lax}{If true then local 3-axis markers will be rendered for every component (default F)}

\item{quiet}{A boolean value that supresses execution feedback output (default T)}

\item{wireframe}{If true object will be rendered as wireframe, otherwise it is shaded (default F)}
}
\description{
Render a compObj object in rgl
}

