% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbToStringColor.R
\name{rgbToStringColor}
\alias{rgbToStringColor}
\title{Converts a 3 vector of rgb values (between 0 and 1) to a string color}
\usage{
rgbToStringColor(rvek, gvek, bvek)
}
\arguments{
\item{rvek}{vector of red values between 0 and 1}

\item{gvek}{vector of green values between 0 and 1}

\item{bvek}{vector of red values between 0 and 1}
}
\description{
Converts a 3 vector of rgb values (between 0 and 1) to a string color
}
\examples{
rgbToStringColor(1,0,0)
# returns [1] "#ff0000"
rgbToStringColor(c(1,0),c(1,0),c(1,0))
# returns [1] "#ffffff" "#000000"
}

