% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw3dlib.R
\docType{package}
\name{mw3dlib}
\alias{mw3dlib}
\alias{mw3dlib-package}
\title{nw3dlib: A package for computating the notorious bar statistic.}
\description{
The mw3dlib package provides utility functions for rgl:
}
\section{mw3dlib functions}{

The mw3dlib functions ...
    addAxesToRgl()
    readCompObjCrazyflie()
    readCompObj()
    renderCompObj()
    rgbToStringColor()
}

